fix:actualizacion de nombres de variables
