--- plataforma/usuario/signals.py
@@ -1,13 +1,0 @@
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-from django.contrib.auth.models import Group
-from django.conf import settings
-from django.contrib.auth import get_user_model
-
-User = get_user_model()
-
-@receiver(post_save, sender=settings.AUTH_USER_MODEL)
-def add_user_to_default_group(sender, instance, created, **kwargs):
-    if created:
-        group, created = Group.objects.get_or_create(name='Participante')
-        instance.groups.add(group)
